remove globals and move to different file
